+++ source/components/expand-button/index.js
@@ -1,1 +1,0 @@
-import "./expand-button.less"