+++ source/components/cloak/index.js
@@ -1,0 +1,1 @@
+import "./cloak.less"