+++ CHANGELOG.md
@@ -12,7 +12,7 @@
 + [x] Update `accordion` for themes
 + [x] Update `calendar` for themes
 + [x] Update `button-group` for themes
-+ [x] Update `checkbox` for themes
++ [x] Update `checkbox` for themes, `style2` removed
 + [x] Update `cards` for themes
 + [x] Added class `error`
 + [x] Updated `breadcrumbs default`, added new style `breadcrumbs arrow`